Interface Changes and Design Changes along w/ Animations
Added two types of animations, design pattern animations which is a
sliding transition between views in app depending on whether transition
is going up or down the heirarchy. And the other is a visual effect
animation, or a splash screen, that occurs at bootup time. Also
reformatted the Home Screen for improved UI and not as cluttered.
Seperated all actions into three tabs (Lost, Found, and Misc) and is
open to suggestions/improvement as well.